fetchAndUpdateCard only needed in steam-getCards
